Add files via upload
Cleaned, visualized, and analyzed bike data with Excel. 
The objective was to showcase an interactive bike sales dashboard where you can compare buyers and non buyers. The reason for this is to show what trends are in people that buy a bike vs. people who don't. In order to do this, the questions I asked about the population (buyers and non-buyers), where in relation to the a) Average income per purchase, b) Customer Age, and c) Customer commute. Now, I added filters in relation to the question such as marital status, region, and level of education to get a better understanding of the population. 

Skills: Data Cleaning, Pivot Tables, Conditional Formatting, Formulas, Interactive dashboards, XLOOKUP.